let you edit the current task
If a task comes from an external server (e.g. Jira), intellij won't let
you edit its summary. If you try, it'll just instantly revert it back to
the original ticket. This lets you override that for the sake of the
intellij-current-task output.

The format is a bit verbose, so as to be robust to any other data I may
want to throw in there in the future:

  {
    "tasks": {
      "JIRA-1234": {
        "description": "my task description"
      }
    }
  }

This currently doesn't get cleaned up ever, but I could definitely
imagine doing that later (e.g. if a task is gone from the idea
workspace, delete it from this as well)